+++ NEWS.md
@@ -1,3 +1,6 @@
+# CHANGES IN litedown VERSION 0.11
+
+
 # CHANGES IN litedown VERSION 0.10
 
 - Added minimal book and website project templates that can be discovered by RStudio from `File` > `New Project` > `New Directory` > `Project using litedown`. You can also create them programmatically with `litedown:::proj_skeleton(path, type)`, where `type` is `"book"` or `"site"` (thanks, @jangorecki, #97).